+++ tomcat/src/main/java/org/apache/coyote/controller/DefaultController.java
@@ -1,10 +1,10 @@
-package org.apache.coyote.handler;
+package org.apache.coyote.controller;
 
 import org.apache.coyote.request.Request;
 import org.apache.coyote.response.HttpStatus;
 import org.apache.coyote.response.Response;
 
-public class DefaultHandler {
+public class DefaultController {
 
     public void response(Request request, Response response) {
         String body = "Hello world!";
