+++ tomcat/src/main/java/org/apache/coyote/controller/ResourceController.java
@@ -1,11 +1,11 @@
-package org.apache.coyote.handler;
+package org.apache.coyote.controller;
 
 import org.apache.coyote.FileReader.FileReader;
 import org.apache.coyote.request.Request;
 import org.apache.coyote.response.HttpStatus;
 import org.apache.coyote.response.Response;
 
-public class ResourceHandler {
+public class ResourceController {
 
     public void getResource(Request request, Response response) {
         String path = request.getPath();
